requirements.txt
streamlit==1.49.1
pandas==2.3.1
pandasql
python-dotenv==1.2.2
groq==0.37.1
semantic-router[local]==0.0.20
chromadb==1.5.5
sentence_transformers==5.2.3